第二次修改
修改了1763三种不同算法的变量类型，通过网站编译

+++ .github/2001/Two/1763_B.c
@@ -2,10 +2,10 @@
 
 int main(void)
 {
-    float a, b;
-    float sum;
+    double a, b;
+    double sum;
 
-    while (scanf("%f", &a) != EOF)
+    while (scanf("%lf", &a) != EOF)
     {
         sum = 0;
         b = a;
@@ -81,7 +81,7 @@ int main(void)
 
         sum = b - sum;
 
-        printf("%.2f\n", sum);    
+        printf("%.2lf\n", sum);    
 
     }
